pa3.3只剩仙剑了，还没找到资源

+++ navy-apps/apps/menu/menu.cpp
@@ -120,7 +120,10 @@ int main(int argc, char *argv[], char *envp[]) {
 
 static void draw_ch(BDF_Font *font, int x, int y, char ch, uint32_t fg, uint32_t bg) {
   SDL_Surface *s = BDF_CreateSurface(font, ch, fg, bg);
-  SDL_Rect dstrect = { .x = x, .y = y };
+  SDL_Rect dstrect = { 
+    .x = static_cast<int16_t>(x),  // 显式转换为 int16_t
+    .y = static_cast<int16_t>(y) 
+  };
   SDL_BlitSurface(s, NULL, screen, &dstrect);
   SDL_FreeSurface(s);
 }
@@ -141,7 +144,10 @@ static void draw_text_row(char *s, int r) {
 
 static void display_menu(int n) {
   clear_display();
-  SDL_Rect rect = { .x = screen->w - logo_sf->w, .y = 0 };
+  SDL_Rect rect = { 
+    .x = static_cast<int16_t>(screen->w - logo_sf->w), 
+    .y = static_cast<int16_t>(0)  // 0 可省略转换，但统一写法更规范
+  };
   SDL_BlitSurface(logo_sf, NULL, screen, &rect);
   printf("Available applications:\n");
   char buf[80];
